sort names alphetically. rename to yo_NG from yo
