--- src/locales/yo/person/first_name.ts
@@ -1,65 +1,0 @@
-export default [
-  'Busayo',
-  'FolaFoluwa',
-  'Ayomide',
-  'Ayomikun',
-  'Temitope',
-  'Abiodun',
-  'Olamide',
-  'Kehinde',
-  'Shola',
-  'Taiwo',
-  'Pelumi',
-  'Olatunbosun',
-  'Iyanuoluwa',
-  'Damilola',
-  'Damola',
-  'Oluwaseun',
-  'Opeyemi',
-  'Oluwatosin',
-  'Toluwalope',
-  'Oluwaseyi',
-  'Abimbola',
-  'Olayinka',
-  'Oluwatobiloba',
-  'Olusola',
-  'Tejumola',
-  'Ayotude',
-  'Adeola',
-  'Abiola',
-  'Idowu',
-  'Adebola',
-  'Mobolaji',
-  'Erioluwa',
-  'Oladara',
-  'Ifeoluwa',
-  'Darasimi',
-  'Adebankole',
-  'Ade',
-  'Adedaramola',
-  'Adedoyin',
-  'Alaba',
-  'Ara',
-  'Ebunoluwa',
-  'Ibukunoluwa',
-  'Imoleoluwa',
-  'Iremide',
-  'Itunuoluwa',
-  'Oluwatumise',
-  'Iyinoluwa',
-  'Omotolani',
-  'Odunayo',
-  'Okiki',
-  'Ola',
-  'Olaoluwa',
-  'Olufeyisayo',
-  'Oluwasemilore',
-  'Oluwatimilehin',
-  'Oluwatofunmi',
-  'Olutomiwa',
-  'Oluwatoyosi',
-  'Omolola',
-  'Opemipo',
-  'Opeoluwa',
-  'Oyindamola',
-];
