--- src/locales/yo/person/last_name.ts
@@ -1,100 +1,0 @@
-export default [
-  'Adegoke',
-  'Ibikunle',
-  'Oparinde',
-  'Oyinlola',
-  'Babtunde',
-  'Afolabi',
-  'Adekunle',
-  'Balogun',
-  'Osuntokun',
-  'Faleti',
-  'Adeyemo',
-  'Soyinka',
-  'Ilori',
-  'Iyanda',
-  'Bamidele',
-  'Aloba',
-  'Owolabi',
-  'Owoyemi',
-  'Aderibigbe',
-  'Kolade',
-  'Adelusi',
-  'Makinde',
-  'Makinwa',
-  'Ojo',
-  'Bakare',
-  'Awolowo',
-  'Awodiran',
-  'Oyadiran',
-  'Oyeniyi',
-  'Omojola',
-  'Salako',
-  'Ajayi',
-  'Jaiyeola',
-  'Ajewole',
-  'Adisa',
-  'Fagunwa',
-  'Ogunbiyi',
-  'Olujimi',
-  'Sobowale',
-  'Afonja',
-  'Afolayan',
-  'Paimo',
-  'Omidina',
-  'Ogindan',
-  'Kosoko',
-  'Bamiloye',
-  'Salami',
-  'Olaiya',
-  'Adeleke',
-  'Owokoniran',
-  'Edun',
-  'Ninalowo',
-  'Oyetoro',
-  'Alabi',
-  'Omotoso',
-  'Adesiyan',
-  'Adegbite',
-  'Ajao',
-  'Olusanya',
-  'Oloyede',
-  'Olabode',
-  'Fajimi',
-  'Adebisi',
-  'Ilupeju',
-  'Fadipe',
-  'Koya',
-  'Oshin',
-  'Odunsi',
-  'Ayoola',
-  'Faloye',
-  'Olowokeere',
-  'Olukunga',
-  'Shekoni',
-  'Odetola',
-  'Oginni',
-  'Ogunde',
-  'Ogunbamigbe',
-  'Olukotun',
-  'Ayandokun',
-  'Olubode',
-  'Falabi',
-  'Ogunyeye',
-  'Olugbayila',
-  'Ademiluyi',
-  'Ogunbo',
-  'Oladipupo',
-  'Oluwagbemi',
-  'Akinrinola',
-  'Oyeyemi',
-  'Aderogba',
-  'Oyaifo',
-  'Ogulana',
-  'Ogunwobi',
-  'Olasupo',
-  'Fasasi',
-  'Morawo',
-  'Adelakun',
-  'Ojua',
-];
