--- src/locales/yo/person/male_first_name.ts
@@ -1,88 +1,0 @@
-export default [
-  'Adebowale',
-  'Adedeji',
-  'Oladimeji',
-  'Adekunle',
-  'Anjolaoluwa',
-  'Akinwale',
-  'Ayodele',
-  'Babawale',
-  'ObaniJesu',
-  'Adekola',
-  'Ogooluwa',
-  'Adeshina',
-  'Abayomi',
-  'Adedayo',
-  'Ikeoluwa',
-  'Oluwamumibori',
-  'Omobobola',
-  'Adetayo',
-  'Esupofo',
-  'Ige',
-  'Folu',
-  'Tade',
-  'Oluwamuyiwa',
-  'Ibidapo',
-  'Adigun',
-  'Lere',
-  'Leke',
-  'Gbadebo',
-  'Oluwagbemiga',
-  'Keji',
-  'Juwon',
-  'Kolawole',
-  'Damife',
-  'Demilade',
-  'Temidayo',
-  'Oluwole',
-  'Oluwafemi',
-  'Oluwasegun',
-  'Abiodun',
-  'Adeniyi',
-  'Adeyi',
-  'Alamu',
-  'Korede',
-  'Ayodeji',
-  'Oluwasanmi',
-  'Olanrewaju',
-  'Olansile',
-  'Ishola',
-  'Akanmu',
-  'Ajani',
-  'Ajala',
-  'Abiona',
-  'Tokunbo',
-  'Ayinde',
-  'Gbolahan',
-  'Babasola',
-  'Adebiyi',
-  'Oke',
-  'Oba',
-  'Inioluwa',
-  'Iseoluwa',
-  'Aderopo',
-  'Seye',
-  'Sijuade',
-  'Shekoni',
-  'Omotayo',
-  'Tomori',
-  'Niyilolawa',
-  'Olumorotimi',
-  'Toki',
-  'Olakunle',
-  'Oladare',
-  'Adekitan',
-  'Adeleke',
-  'Osunleke',
-  'Adeolu',
-  'Adeoti',
-  'Adesoji',
-  'Afolarin',
-  'Akinlabi',
-  'Akinkunmi',
-  'Alade',
-  'Babatunji',
-  'Durodola',
-  'Ekundayo',
-  'Gbowoade',
-];
