+++ src/locales/yo_NG/person/female_first_name.ts
@@ -1,87 +1,86 @@
 export default [
   'Aanuoluwapo',
-  'Abosede',
-  'Abeni',
   'Abebi',
+  'Abeni',
+  'Abosede',
+  'Adebukola',
   'Adenike',
   'Adepeju',
-  'Adunni',
-  'Adebukola',
   'Adesewa',
+  'Adesua',
+  'Adetoke',
+  'Adetoun',
+  'Adunni',
+  'Ajoke',
   'Amoke',
   'Amope',
   'Arike',
-  'Asake',
   'Arinola',
+  'Asake',
+  'Atinuke',
   'Awero',
   'Ayinke',
-  'Ajoke',
   'Ayoka',
-  'Boluwatife',
   'Bolatito',
+  'Boluwatife',
   'Bunmi',
   'Doyinsola',
-  'Omobolanle',
-  'Erioluwa',
-  'Ewatomi',
   'Eniola',
+  'Ewatomi',
   'Fadekemi',
-  'Fikayomi',
-  'Fibikemi',
+  'Faderera',
   'Fehintola',
+  'Fibikemi',
+  'Fikayomi',
   'Folashade',
   'Ibironke',
   'Iretioluwa',
   'Iyabode',
   'Iyadunni',
   'Kikelomo',
-  'Ololade',
   'Modupe',
-  'Moyosore',
-  'Morenike',
   'Mofifoluwa',
   'Mojisola',
-  'Morayo',
-  'Mosinmileoluwa',
   'Mojisoluwa',
-  'Motunrayo',
-  'Mosunmola',
-  'Morolake',
   'Moradeke',
+  'Morayo',
+  'Morenike',
+  'Morolake',
+  'Mosinmileoluwa',
+  'Mosunmola',
+  'Motunrayo',
+  'Moyosore',
   'Ninioluwa',
   'Olajumoke',
   'Olasunmbo',
-  'Oluwakemi',
+  'Ololade',
   'Olufunke',
   'Olufunmilayo',
+  'Oluwakemi',
+  'Omobolanle',
   'Omodunni',
   'Omolabake',
   'Omolara',
   'Omosalewa',
   'Omotara',
   'Omotola',
-  'Oreofe',
   'Omotoun',
-  'Ronke',
+  'Omowumi',
+  'Oreofe',
+  'Oyenike',
+  'Oyindasola',
   'Radeke',
+  'Ronke',
   'Segilola',
-  'Sowande',
-  'Subomi',
   'Similoluwa',
   'Simisola',
+  'Sowande',
+  'Subomi',
   'Titilayo',
-  'Toluwanimi',
   'Tolulope',
+  'Toluwanimi',
   'Wuraola',
+  'Yejide',
   'Yetunde',
   'Yewande',
-  'Oyenike',
-  'Adetoke',
-  'Omowumi',
-  'Adetoun',
-  'Atinuke',
-  'Oyindasola',
-  'Yejide',
-  'Adesua',
-  'Faderera',
 ];
