+++ src/locales/yo_NG/person/first_name.ts
@@ -1,0 +1,65 @@
+export default [
+  'Abimbola',
+  'Abiodun',
+  'Abiola',
+  'Ade',
+  'Adebankole',
+  'Adebola',
+  'Adedaramola',
+  'Adedoyin',
+  'Adeola',
+  'Alaba',
+  'Ara',
+  'Ayomide',
+  'Ayomikun',
+  'Ayotude',
+  'Busayo',
+  'Damilola',
+  'Damola',
+  'Darasimi',
+  'Ebunoluwa',
+  'Erioluwa',
+  'FolaFoluwa',
+  'Ibukunoluwa',
+  'Idowu',
+  'Ifeoluwa',
+  'Imoleoluwa',
+  'Iremide',
+  'Itunuoluwa',
+  'Iyanuoluwa',
+  'Iyinoluwa',
+  'Kehinde',
+  'Mobolaji',
+  'Odunayo',
+  'Okiki',
+  'Ola',
+  'Oladara',
+  'Olamide',
+  'Olaoluwa',
+  'Olatunbosun',
+  'Olayinka',
+  'Olufeyisayo',
+  'Olusola',
+  'Olutomiwa',
+  'Oluwasemilore',
+  'Oluwaseun',
+  'Oluwaseyi',
+  'Oluwatimilehin',
+  'Oluwatobiloba',
+  'Oluwatofunmi',
+  'Oluwatosin',
+  'Oluwatoyosi',
+  'Oluwatumise',
+  'Omolola',
+  'Omotolani',
+  'Opemipo',
+  'Opeoluwa',
+  'Opeyemi',
+  'Oyindamola',
+  'Pelumi',
+  'Shola',
+  'Taiwo',
+  'Tejumola',
+  'Temitope',
+  'Toluwalope',
+];
