+++ src/locales/yo_NG/person/index.ts
