+++ src/locales/yo_NG/person/last_name.ts
@@ -1,0 +1,100 @@
+export default [
+  'Adebisi',
+  'Adegbite',
+  'Adegoke',
+  'Adekunle',
+  'Adelakun',
+  'Adeleke',
+  'Adelusi',
+  'Ademiluyi',
+  'Aderibigbe',
+  'Aderogba',
+  'Adesiyan',
+  'Adeyemo',
+  'Adisa',
+  'Afolabi',
+  'Afolayan',
+  'Afonja',
+  'Ajao',
+  'Ajayi',
+  'Ajewole',
+  'Akinrinola',
+  'Alabi',
+  'Aloba',
+  'Awodiran',
+  'Awolowo',
+  'Ayandokun',
+  'Ayoola',
+  'Babtunde',
+  'Bakare',
+  'Balogun',
+  'Bamidele',
+  'Bamiloye',
+  'Edun',
+  'Fadipe',
+  'Fagunwa',
+  'Fajimi',
+  'Falabi',
+  'Faleti',
+  'Faloye',
+  'Fasasi',
+  'Ibikunle',
+  'Ilori',
+  'Ilupeju',
+  'Iyanda',
+  'Jaiyeola',
+  'Kolade',
+  'Kosoko',
+  'Koya',
+  'Makinde',
+  'Makinwa',
+  'Morawo',
+  'Ninalowo',
+  'Odetola',
+  'Odunsi',
+  'Ogindan',
+  'Oginni',
+  'Ogulana',
+  'Ogunbamigbe',
+  'Ogunbiyi',
+  'Ogunbo',
+  'Ogunde',
+  'Ogunwobi',
+  'Ogunyeye',
+  'Ojo',
+  'Ojua',
+  'Olabode',
+  'Oladipupo',
+  'Olaiya',
+  'Olasupo',
+  'Olowokeere',
+  'Oloyede',
+  'Olubode',
+  'Olugbayila',
+  'Olujimi',
+  'Olukotun',
+  'Olukunga',
+  'Olusanya',
+  'Oluwagbemi',
+  'Omidina',
+  'Omojola',
+  'Omotoso',
+  'Oparinde',
+  'Oshin',
+  'Osuntokun',
+  'Owokoniran',
+  'Owolabi',
+  'Owoyemi',
+  'Oyadiran',
+  'Oyaifo',
+  'Oyeniyi',
+  'Oyetoro',
+  'Oyeyemi',
+  'Oyinlola',
+  'Paimo',
+  'Salako',
+  'Salami',
+  'Shekoni',
+  'Sobowale',
+  'Soyinka',
+];
