+++ src/locales/yo_NG/person/last_name_pattern.ts
