+++ src/locales/yo_NG/person/male_first_name.ts
@@ -1,0 +1,88 @@
+export default [
+  'Abayomi',
+  'Abiodun',
+  'Abiona',
+  'Adebiyi',
+  'Adebowale',
+  'Adedayo',
+  'Adedeji',
+  'Adekitan',
+  'Adekola',
+  'Adekunle',
+  'Adeleke',
+  'Adeniyi',
+  'Adeolu',
+  'Adeoti',
+  'Aderopo',
+  'Adeshina',
+  'Adesoji',
+  'Adetayo',
+  'Adeyi',
+  'Adigun',
+  'Afolarin',
+  'Ajala',
+  'Ajani',
+  'Akanmu',
+  'Akinkunmi',
+  'Akinlabi',
+  'Akinwale',
+  'Alade',
+  'Alamu',
+  'Anjolaoluwa',
+  'Ayinde',
+  'Ayodeji',
+  'Ayodele',
+  'Babasola',
+  'Babatunji',
+  'Babawale',
+  'Damife',
+  'Demilade',
+  'Durodola',
+  'Ekundayo',
+  'Esupofo',
+  'Folu',
+  'Gbadebo',
+  'Gbolahan',
+  'Gbowoade',
+  'Ibidapo',
+  'Ige',
+  'Ikeoluwa',
+  'Inioluwa',
+  'Iseoluwa',
+  'Ishola',
+  'Juwon',
+  'Keji',
+  'Kolawole',
+  'Korede',
+  'Leke',
+  'Lere',
+  'Niyilolawa',
+  'Oba',
+  'ObaniJesu',
+  'Ogooluwa',
+  'Oke',
+  'Oladare',
+  'Oladimeji',
+  'Olakunle',
+  'Olanrewaju',
+  'Olansile',
+  'Olumorotimi',
+  'Oluwafemi',
+  'Oluwagbemiga',
+  'Oluwamumibori',
+  'Oluwamuyiwa',
+  'Oluwasanmi',
+  'Oluwasegun',
+  'Oluwole',
+  'Omobobola',
+  'Omotayo',
+  'Osunleke',
+  'Seye',
+  'Shekoni',
+  'Sijuade',
+  'Tade',
+  'Temidayo',
+  'Toki',
+  'Tokunbo',
+  'Tomori',
+];
